+++ diesel_derives/src/attrs.rs
@@ -33,6 +33,7 @@ pub struct AttributeSpanWrapper<T> {
 pub enum FieldAttr {
     Embed(Ident),
     SkipInsertion(Ident),
+    SkipUpdate(Ident),
 
     ColumnName(Ident, SqlIdentifier),
     SqlType(Ident, TypePath),
@@ -145,6 +146,7 @@ impl Parse for FieldAttr {
         match &*name_str {
             "embed" => Ok(FieldAttr::Embed(name)),
             "skip_insertion" => Ok(FieldAttr::SkipInsertion(name)),
+            "skip_update" => Ok(FieldAttr::SkipUpdate(name)),
 
             "column_name" => Ok(FieldAttr::ColumnName(
                 name,
@@ -199,6 +201,7 @@ impl MySpanned for FieldAttr {
         match self {
             FieldAttr::Embed(ident)
             | FieldAttr::SkipInsertion(ident)
+            | FieldAttr::SkipUpdate(ident)
             | FieldAttr::ColumnName(ident, _)
             | FieldAttr::SqlType(ident, _)
             | FieldAttr::TreatNoneAsNull(ident, _)